Refactor: Shifted Pages & Components to Index.js
