+++ apps/frontend/src/app/assets/constants/Config/DashboardConfig/index.js
