+++ apps/frontend/src/app/assets/constants/Config/EChartConfig/index.js
