+++ apps/frontend/src/app/assets/constants/Config/LineChartConfig/index.js
