+++ apps/frontend/src/app/components/Footer/index.js
