+++ apps/frontend/src/app/components/Icons/index.js
