+++ apps/frontend/src/app/components/Interaction/index.js
