+++ apps/frontend/src/app/pages/RecentActivities/index.js
